+++ sdk/communication/Azure.Communication.Chat/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Release History
 
 ## 1.1.0-beta.1 (Unreleased)
+### Added
+- Added support for metadata in messages.
+- Added new overloads for sendmessage and updatemessage operations with options : SendChatMessageOptions and UpdateChatMessageOptions. 
 
 
 ## 1.0.1 (2021-05-11)