+++ sdk/textanalytics/Azure.AI.TextAnalytics/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Release History
 
-## 5.1.0-beta.7 (Unreleased)
+## 5.1.0-beta.8 (Unreleased)
+### Breaking changes
+- Renamed `StartAnalyzeBatchActions` to `StartAnalyzeActions`.
+- Renamed `AnalyzeBatchActionsOperation` to `AnalyzeActionsOperation`.
+- Renamed `AnalyzeBatchActionsResult` to `AnalyzeActionsResult`.
+- Renamed `AnalyzeBatchActionsOptions` to `AnalyzeActionsOptions`.
+- `TextAnalyticsActions` now takes `xxAction` types, instead of `xxOptions` types. Renames and types are as follow:
+  - `ExtractKeyPhrasesOptions` changed to new type `ExtractKeyPhrasesActions`.
+  - `RecognizeEntitiesOptions` changed to new type `RecognizeEntitiesActions`.
+  - `RecognizePiiEntitiesOptions` changed to new type `RecognizePiiEntitiesActions`.
+  - `RecognizeLinkedEntitiesOptions` changed to new type `RecognizeLinkedEntitiesActions`.
+  - `AnalyzeSentimentOptions` changed to new type `AnalyzeSentimentActions`.
+- Renamed type `TextAnalyticsActionDetails` to `TextAnalyticsActionResult`.
+
+## 5.1.0-beta.7 (2021-05-18)
 ### New features
 - Added property `DisableServiceLogs` to `TextAnalyticsRequestOptions`.
+- Added support for Sentiment Analysis as an action type for `StartAnalyzeBatchActions`.
+- Changed type of `IncludeOpinionMining` to `bool?`.
 
 ### Breaking changes
 - The client defaults to the latest supported service version, which currently is `3.1-preview.5`.